np.int and np.float are deprecated
https://numpy.org/devdocs/release/1.20.0-notes.html#using-the-aliases-of-builtin-types-like-np-int-is-deprecated

+++ tests/test_approval.py
@@ -152,7 +152,7 @@ def combined_approval_ballots(min_cands=1, max_cands=25, min_voters=1,
     """
     n_cands = integers(min_value=min_cands, max_value=max_cands)
     n_voters = integers(min_value=min_voters, max_value=max_voters)
-    return arrays(np.int, tuples(n_voters, n_cands), elements=integers(-1, 1))
+    return arrays(int, tuples(n_voters, n_cands), elements=integers(-1, 1))
 
 
 @pytest.mark.parametrize("tiebreaker", ['random', 'order'])
