+++ tests/test_utility_winner.py
@@ -76,7 +76,7 @@ def random_utilities(min_cands=1, max_cands=25, min_voters=1, max_voters=100):
     """
     n_cands = integers(min_value=min_cands, max_value=max_cands)
     n_voters = integers(min_value=min_voters, max_value=max_voters)
-    return arrays(np.float, tuples(n_voters, n_cands), elements=floats(0, 1))
+    return arrays(float, tuples(n_voters, n_cands), elements=floats(0, 1))
 
 
 @pytest.mark.parametrize("tiebreaker", ['random', 'order'])
